+++ meta.rkt
@@ -12,50 +12,6 @@
              numlock
              wheel-up wheel-down wheel-left wheel-right))
 
-(define racket-builtin-form*
-  (list
-   ;;; define
-   "(define )"
-   "(define () )"
-   ;;; let
-   "(let ([]) )"
-   "(let* ([]) )"
-   "(letrec ([]) )"
-   ;;; λ
-   "(lambda () )"
-   "(λ () )"
-   ;;; branch
-   "(cond
-  [else ])"
-   "(match 
-  [else ])"
-   ;;; module
-   "(require )"
-   "(provide )"
-   "(all-defined-out)"
-   "(only-in )"
-   "(except-in )"
-   "(prefix-in )"
-   "(rename-in )"
-   "(combine-in )"
-   "(relative-in )"
-   "(only-meta-in )"
-   "(for-syntax )"
-   "contract-out"
-   ;;; loop
-   "(for ([])
-  )"
-   "(for/list ([])
-  )"
-   "(for/vector ([])
-  )"
-   "(for/hash ([])
-  )"
-   "(for/hasheq ([])
-  )"
-   "(for/hasheqv ([])
-  )"))
-
 (define latex-complete
   #hash(("\\" . "\\")
         ;;; logic